requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
requests
python-dotenv
uvicorn[standard]==0.15.0